NLP-Service/requirements.txt
numpy
torch
nltk
pyspellchecker
fuzzywuzzy
python-Levenshtein
flask
flask-cors